+++ react/__mocks__/vtex.device-detector.js
@@ -1,3 +1,11 @@
+import React from 'react'
+
 export const useDevice = () => {
   return { isMobile: true }
+}
+
+export const withDevice = WrappedComponent => {
+  return function WrappedWithDevice(props) {
+    return <WrappedComponent {...props} />
+  }
 }