+++ react/components/SearchResult.js
@@ -125,6 +125,7 @@ class SearchResult extends Component {
       orderBy,
       mobileLayout,
       isMobile,
+      showBreadcrumbsOnMobile,
     } = this.props
     const {
       mobileLayoutMode,
@@ -181,8 +182,8 @@ class SearchResult extends Component {
             this.props.blockClass
           )} w-100 mw9`}
         >
-          {!isMobile && (
-            <div className={styles.breadcrumb}>
+          {(!isMobile || showBreadcrumbsOnMobile) && (
+            <div data-testid="breadcrumb" className={styles.breadcrumb}>
               <ExtensionPoint id="breadcrumb" {...breadcrumbsProps} />
             </div>
           )}